refactor(move): Domain 디렉토리 생성
Domain 디렉토리로 옮긴 파일들:
- BenefitCalculator.js
- Discount.js
- OrderManager.js
- VisitDate.js
